+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.18.0.0
 
+* Add `sumUTxO` and `sumCoinUTxO`
+* Deprecate `balance` and `coinBalance` in favor of `sumUTxO` and `sumCoinUTxO`
 * Remove `delegators` field from JSON serialiser for `DRepState` for correct round-tripping. #5004
 * Change `TxBody` to an associated `data` family
 * Remove `HeapWords` instances for: #5001